Refactoring of test handling. Refs #19.
Use a ConfigParser in stead of paste deploy stuff.
