--- crabpy/__init__.py
@@ -1,3 +1,0 @@
-def main(global_config, **settings):
-    """ This function returns a Pyramid WSGI application.
-    """
