+++ requirements.txt
@@ -1,0 +1,3 @@
+suds-jurko==0.6
+dogpile.cache==0.5.6
+six==1.9.0
