+++ tests/test.ini
@@ -1,9 +1,7 @@
-[app:main]
-use = egg:crabpy
+[crab]
+run_integration_tests = True
 
-crab_run_integration_tests = True
-
-capakey_run_integration_tests = True
-capakey_user = vandaeko
-capakey_password = f77737b9-55cc-4c4a-b86a-895d328c88f4
-
+[capakey]
+run_integration_tests = True
+user = vandaeko
+password = f77737b9-55cc-4c4a-b86a-895d328c88f4
